fix : 파라미터 제거하면서 생긴 버그 수정 & 기존 코드 개선

+++ app/src/main/java/com/sigmai/stylemento/data/model/response/review/Review.kt
@@ -20,7 +20,7 @@ data class Review(
 ) {
     fun toEntity() : Review {
         return Review(
-            rating = star,
+            rating = star.toFloat(),
             nickname = reviewerId,
             profileImageUrl = null,
             tall = height,
